// widgetworld-master/src/app/Interfaces/inventory.ts
export interface ClassificationType {
  description: string;
  id: number;
  name: string;
}

export interface ConstructionType {
  description: string;
  id: number;
  name: string;
}

export interface IlluminationType {
  description: string;
  id: number;
  name: string;
}

export interface Properties {
  base_segment: string;
  eff_freq_avg: number;
  eff_freq_min: number;
  eff_reach_net: number;
  eff_reach_pct: number;
  frame_id: number;
  freq_avg: number;
  imp: number;
  imp_inmkt: number;
  imp_target: number;
  imp_target_inmkt: number;
  index_comp_target: number;
  market: string;
  measures_type: string;
  media_type: string;
  pct_comp_imp_target: number;
  pct_comp_imp_target_inmkt: number;
  pct_comp_pop_target_inmkt: number;
  pct_imp_inmkt: number;
  pct_imp_target_inmkt: number;
  period_days: number;
  plant_frame_id: string;
  plant_operator: string;
  pop_inmkt: number;
  pop_target_inmkt: number;
  reach_net: number;
  reach_pct: number;
  target_geo: string;
  target_segment: string;
  trp: number;
}

export interface InventorySpot {
  classification_type: ClassificationType;
  construction_type: ConstructionType;
  digital: boolean;
  geometry: any[];
  illumination_type: IlluminationType;
  max_height: string;
  max_width: string;
  properties: Properties;
}